+++ src/app/model/menu.interface.ts
@@ -1,0 +1,7 @@
+
+export type MenuItem = {
+    icon: string,
+    label: string,
+    route?: string,
+    subItems?: MenuItem[]
+}
