+++ src/app/pages/layout/layout.component.html
@@ -1,9 +1,7 @@
-<div class="min-h-screen bg-slate-700">
-   <app-custom-sidebar></app-custom-sidebar>
+<div class="min-h-screen bg-neutral-100">
+  <app-custom-sidebar></app-custom-sidebar>
+  <app-navbar></app-navbar>
   <div class="p-4 xl:ml-80">
-    <nav class="block w-full max-w-full bg-transparent text-white shadow-none rounded-xl transition-all px-0 py-1">
-      <app-navbar></app-navbar>
-    </nav>
     <router-outlet></router-outlet>
     <div class="text-blue-gray-600">
       <footer class="py-2">
