+++ src/app/shared/board/board.component.html
@@ -1,17 +1,10 @@
 <div class="min-h-screen">
   <div class="container mx-auto">
     <div class="grid grid-cols-4 lg:grid-cols-4 gap-2">
-      <!-- Projects Section -->
-      <!--  <app-projects-board class="col-span-3 lg:col-span-3"></app-projects-board> -->
-      <!-- Tasks Section -->
-       <app-tasks-board class="col-span-2 lg:col-span-2"></app-tasks-board>
-      <!-- Subtasks Section -->
-       <app-subtasks-board class="col-span-2 lg:col-span-2"></app-subtasks-board>
-      <!-- Issues Section -->
-       <app-issues-board class="col-span-2 lg:col-span-2"></app-issues-board>
-      <!-- Comments Section -->
-       <app-comments-board class="col-span-2 lg:col-span-2"></app-comments-board>
+      <app-tasks-board class="col-span-2 lg:col-span-2"></app-tasks-board>
+      <app-subtasks-board class="col-span-2 lg:col-span-2"></app-subtasks-board>
+      <app-issues-board class="col-span-2 lg:col-span-2"></app-issues-board>
+      <app-comments-board class="col-span-2 lg:col-span-2"></app-comments-board>
     </div>
   </div>
 </div>
-
