+++ src/app/shared/custom-sidebar/custom-sidebar.component.ts
@@ -1,16 +1,11 @@
-import { CommonModule } from '@angular/common';
-import { Component, signal } from '@angular/core';
+import { CommonModule } from '@angular/common'
+import { Component, signal } from '@angular/core'
 import { MatListModule } from '@angular/material/list'
 import { MatIconModule } from '@angular/material/icon'
-import { RouterLink, RouterModule } from '@angular/router';
-import { AuthService } from '../../service/auth/auth.service';
-
-export type MenuItem = {
-  icon: string,
-  label: string,
-  route?: string,
-  subItems?: MenuItem[]
-}
+import { RouterLink, RouterModule } from '@angular/router'
+import { AuthService } from '../../service/auth/auth.service'
+import { MENU_ITEMS } from '../menu-items/menu-items'
+import { MenuItem } from '../../model/menu.interface'
 
 @Component({
   selector: 'app-custom-sidebar',
@@ -23,41 +18,10 @@ export class CustomSidebarComponent {
 
   constructor(private authService: AuthService) { }
 
-  menuItems = signal<MenuItem[]>([
-    {
-      icon: 'dashboard',
-      label: 'Dashboard',
-      route: 'pages/dashboard',
-    },
-    {
-      icon: 'folder_open',
-      label: 'Projects',
-      route: 'pages/projects',
-    },
-    {
-      icon: 'assignment',
-      label: 'Task',
-      route: 'pages/tasks',
-    },
-    {
-      icon: 'subdirectory_arrow_right',
-      label: 'SubTasks',
-      route: 'pages/subtasks',
-    },
-    {
-      icon: 'bug_report',
-      label: 'Issues',
-      route: 'pages/issues',
-    },
-    {
-      icon: 'comment',
-      label: 'Comments',
-      route: 'pages/comments',
-    }
-  ])
+  menuItems = signal<MenuItem[]>(MENU_ITEMS)
 
   onLogout(): void {
-    this.authService.logout();
+    this.authService.logout()
   }
 
 }
