+++ src/app/shared/navbar/navbar.component.html
@@ -1,4 +1,4 @@
-<div class="flex flex-col-reverse justify-between gap-6 md:flex-row md:items-center">
+<!-- <div class="flex flex-col-reverse justify-between gap-6 md:flex-row md:items-center">
   <div class="" >
     <nav class="flex items-center space-x-4 text-gray-600 text-sm">
       <ng-container *ngFor="let breadcrumb of breadcrumbs; let last = last">
@@ -112,4 +112,9 @@
       </li>
     </ul>
   </div>
+</div> -->
+<div class="flex flex-col-reverse justify-between md:flex-row md:items-center bg-white w-full">
+  <nav class="flex justify-center items-center ml-auto text-gray-600 text-sm h-16 w-full md:w-auto m-2">
+    <p-avatar image="https://primefaces.org/cdn/primeng/images/demo/avatar/amyelsner.png" shape="circle" />
+  </nav>
 </div>
